TechWeek_Hackathon/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pandas
google-generativeai
markdown-it-py
